move connected app to service/
