+++ packages/devtools_app/lib/devtools_app.dart
@@ -58,6 +58,8 @@ export 'src/screens/vm_developer/object_inspector/vm_object_model.dart';
 export 'src/screens/vm_developer/vm_developer_tools_controller.dart';
 export 'src/screens/vm_developer/vm_developer_tools_screen.dart';
 export 'src/screens/vm_developer/vm_service_private_extensions.dart';
+export 'src/service/connected_app/connected_app.dart';
+export 'src/service/connected_app/connection_info.dart';
 export 'src/service/editor/editor_client.dart';
 export 'src/service/json_to_service_cache.dart';
 export 'src/service/service_extensions.dart';
@@ -69,8 +71,6 @@ export 'src/service/vm_service_wrapper.dart';
 export 'src/shared/analytics/analytics.dart';
 export 'src/shared/analytics/analytics_controller.dart';
 export 'src/shared/charts/treemap.dart';
-export 'src/shared/connected_app/connected_app.dart';
-export 'src/shared/connected_app/connection_info.dart';
 export 'src/shared/console/console_service.dart';
 export 'src/shared/console/eval/eval_service.dart';
 export 'src/shared/console/eval/inspector_tree.dart';
