+++ packages/devtools_app/lib/src/framework/disconnect_observer.dart
@@ -9,10 +9,10 @@ import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_app_shared/utils.dart';
 import 'package:flutter/material.dart';
 
+import '../service/connected_app/connection_info.dart';
 import '../shared/analytics/analytics.dart' as ga;
 import '../shared/analytics/constants.dart' as gac;
 import '../shared/config_specific/import_export/import_export.dart';
-import '../shared/connected_app/connection_info.dart';
 import '../shared/framework/routing.dart';
 import '../shared/globals.dart';
 import '../shared/primitives/query_parameters.dart';
