+++ packages/devtools_app/lib/src/framework/home_screen.dart
@@ -12,10 +12,10 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
+import '../service/connected_app/connection_info.dart';
 import '../shared/analytics/analytics.dart' as ga;
 import '../shared/analytics/constants.dart' as gac;
 import '../shared/config_specific/import_export/import_export.dart';
-import '../shared/connected_app/connection_info.dart';
 import '../shared/framework/routing.dart';
 import '../shared/framework/screen.dart';
 import '../shared/globals.dart';
