+++ packages/devtools_app/lib/src/service/service_manager.dart
@@ -10,7 +10,6 @@ import 'package:logging/logging.dart';
 import 'package:vm_service/vm_service.dart' hide Error;
 
 import '../shared/analytics/analytics.dart' as ga;
-import '../shared/connected_app/connected_app.dart';
 import '../shared/console/console_service.dart';
 import '../shared/diagnostics/inspector_service.dart';
 import '../shared/globals.dart';
@@ -18,6 +17,7 @@ import '../shared/managers/error_badge_manager.dart';
 import '../shared/server/server.dart' as server;
 import '../shared/title.dart';
 import '../shared/utils/utils.dart';
+import 'connected_app/connected_app.dart';
 import 'service_registrations.dart' as registrations;
 import 'timeline_streams.dart';
 import 'vm_flags.dart';
